Add govuk-frontend scripts and assets
Add JS, CSS, images and fonts from govuk-frontend:
https://github.com/alphagov/govuk-frontend

This will let us add styled GOV.UK components to the frontend, such
as the GOV.UK header.

Copying the assets is not the recommended way of installing the
design system. The recommended approach is to install it with npm,
but we don't have a frontend build pipeline yet so it's easier to
add everything directly for now.