Get new auth code when the user visits the /upload page
Repeat some of the authentication workflow to get a new code when
the user visits /upload.

This results in a redirect to a URL with ?code=abcd in the URL
parameters. This parameter is accessible by the frontend, so the
JavaScript can (in future) convert the code to a token just as it
would in a purely client-side application.