Use UUIDs as upload filenames rather than original filenames
Get a UUID from the database for each file, and use it as the filename
when uploading the file to S3.

This will allow backend file checks and the file export task to link
files in the API with files in S3.

It also means that we don't have to encode file names that contain
characters that aren't allowed in S3.