Replace `Future.apply` with `Future.successful` in controllers
In each case, we already have the data to return but it needs to be
wrapped in a Future. `Future.apply` and `Future.successful` both
wrap data in a future, but `Future.successful` is a better option in
these cases because it wraps the data in a Future synchronously on
the current thread. `Future.apply` starts an asyncronous process,
which is unnecessary unless you are performing another async
operation.