Calculate checksum incrementally
Use the asmCrypto.js library to calculate a file's checksum
incrementally, by loading it into memory in chunks rather than
reading the entire file at once.

This is slower than using SubtleCrypto, which uses low-level
functions built into the browser, but SubtleCrypto needs the full
contents of a file which limits the size of the files that we can
process.