Fix bug which caused polling to carry on too long
There was an error where we were trying to set the inner text
on an element which didn't support it. This meant that the next
line which stopped the polling wasn't working. There's no reason
to update this div so I've deleted this line.

There was second issue where the javascript wasn't updating the
error page correctly if there were virus or checksum errors. I've
fixed this in the javascript and tidied up the .scala.html file.