Temporarily revert WebAssembly progress monitor
Remove the console logs which happen when the WebAssembly checksum
calculator reports a progress update.

The version of the WebAssembly library which supports the `progress`
argument (version 0.1.5) has not yet been published to npm, so the
local npm build is failing. Removing the `progress` argument means
we can use the latest npm version (0.1.4).