Add temporary WebAssembly override for checksum calculation
Check a value in local storage to decide whether to use WebAssembly
or the JS fallback when calculating a file's checksum.

This won't change the default behaviour (so WASM will be used when
it's available) but it will let us test the performance of both
approaches without having to redeploy the app.