Fetch JWT access token rather than opaque token
Add an `audience` parameter to the frontend token query,
which makes Auth0 return a JWT access token which contains
user claims, so you don't have to make a separate request
for the user profile.

This should mean that we can send the JWT access token to
the API, and it can verify the token without an extra request.