Make uploads of client-side metadata batches sequential
Change the way the frontend sends client-side metadata to the API.
Send the batches sequentially rather than in parallel. The parallel
requests have been returning 504 errors since we added an index to
the file path column in the DB.

These API requests are much quicker than the checksum and file upload
steps, so it shouldn't cause a performance problem if we send them
sequentially.