Remove IE8 version of govuk-frontend
The styles were being loaded in Chrome for Mac, and it was breaking the layout:
the main content was being left-aligned rather than centred.

We cannot test this prototype in IE anyway because you cannot upload directories
in IE, so it's safe to remove the IE version of govuk-frontend. We can revisit
this issue in Beta when we add IE zip-upload support.